requirements.txt
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
cffi==2.0.0
cryptography==46.0.5
Django==6.0.2
django-environ==0.13.0
django-mongodb-backend @ https://github.com/aclark4life/django-mongodb-backend/archive/refs/heads/INTPYTHON-527.zip#sha256=84f44a97498ab179a99d062ce86ad3c8ead256118ec0b2b0ee87476f3fd18aeb
djangorestframework==3.16.1
dnspython==2.8.0
drf-spectacular==0.29.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
packaging==26.0
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
pymongo==4.16.0
pymongo-auth-aws==1.3.0
pymongocrypt==1.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
